requirements.txt
ChatterBot>=1.0.5
chatterbot-corpus>=1.2.0
click>=6.7
Flask>=1.1.0
future>=0.16.0
graypy>=0.2.14
itsdangerous>=0.24
Jinja2>=2.10
jsondatabase>=0.1.7
MarkupSafe>=1.0
nltk>=3.4.5
oauthlib>=2.0.1
pymongo>=3.4.0
python-twitter>=3.2.1
requests>=2.20.0
requests-oauthlib>=0.8.0
six>=1.10.0
Werkzeug>=0.15.3
gensim>=3.8.1
konlpy>=0.5.2